requirements.txt
anyio==3.6.1
asgiref==3.5.2
attrs==22.1.0
backoff==2.1.2
backports.cached-property==1.0.2
black==22.3.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
Deprecated==1.2.13
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.79.0
googleapis-common-protos==1.56.4
graphql-core==3.2.1
greenlet==2.0.0a2
grpcio==1.48.0rc1
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
mangum==0.15.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
opentelemetry-api==1.12.0rc2
opentelemetry-exporter-otlp==1.12.0rc2
opentelemetry-exporter-otlp-proto-grpc==1.12.0rc2
opentelemetry-exporter-otlp-proto-http==1.12.0rc2
opentelemetry-instrumentation==0.32b0
opentelemetry-instrumentation-asgi==0.32b0
opentelemetry-instrumentation-aws-lambda==0.32b0
opentelemetry-instrumentation-dbapi==0.32b0
opentelemetry-instrumentation-fastapi==0.32b0
opentelemetry-instrumentation-grpc==0.32b0
opentelemetry-instrumentation-requests==0.32b0
opentelemetry-instrumentation-sqlalchemy==0.32b0
opentelemetry-proto==1.12.0rc2
opentelemetry-sdk==1.12.0rc2
opentelemetry-semantic-conventions==0.32b0
opentelemetry-util-http==0.32b0
orjson==3.7.11
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
protobuf==3.20.1
psycopg2-binary==2.9.3
py==1.11.0
pydantic==1.9.1
Pygments==2.12.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
strawberry-graphql==0.124.0.dev1659690947
tomli==2.0.1
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.11
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
